+++ relayer-cli/src/commands/create.rs
@@ -3,6 +3,7 @@ use abscissa_core::{Command, Help, Options, Runnable};
 
 use crate::commands::create::channel::CreateChannelCommand;
 use crate::commands::create::connection::CreateConnectionCommand;
+use crate::commands::tx::client::TxCreateClientCmd;
 
 mod channel;
 mod connection;
@@ -13,6 +14,10 @@ pub enum CreateCmds {
     /// Generic `help`
     #[options(help = "Get usage information")]
     Help(Help<Self>),
+
+    /// Subcommand for creating a `client`
+    #[options(help = "Create a new IBC client")]
+    Client(TxCreateClientCmd),
 
     /// Subcommand for creating a `connection`
     #[options(help = "Create a new connection between two chains")]