+++ relayer/src/chain.rs
@@ -256,6 +256,7 @@ pub trait Chain: Sized {
 
     fn build_header(
         &self,
+        trusted_height: ICSHeight,
         trusted_light_block: Self::LightBlock,
         target_light_block: Self::LightBlock,
     ) -> Result<Self::Header, Error>;