+++ relayer/src/chain/mock.rs
@@ -307,16 +307,14 @@ impl Chain for MockChain {
 
     fn build_header(
         &self,
+        trusted_height: Height,
         trusted_light_block: Self::LightBlock,
         target_light_block: Self::LightBlock,
     ) -> Result<Self::Header, Error> {
         Ok(Self::Header {
             signed_header: target_light_block.signed_header.clone(),
             validator_set: target_light_block.validators,
-            trusted_height: Height::new(
-                self.id().version(),
-                u64::from(trusted_light_block.signed_header.header.height),
-            ),
+            trusted_height,
             trusted_validator_set: trusted_light_block.validators,
         })
     }
@@ -349,6 +347,8 @@ pub mod test_utils {
             key_name: "".to_string(),
             store_prefix: "".to_string(),
             gas: None,
+            fee_amount: Some(1000),
+            fee_denom: "stake".to_string(),
             max_msg_num: None,
             max_tx_size: None,
             clock_drift: Duration::from_secs(5),